Move fold functions and type to Cardano.Api.LedgerState.Fold
