+++ cardano-api/internal/Cardano/Api/LedgerState.hs
@@ -73,3 +73,4 @@ module Cardano.Api.LedgerState
   where
 
 import           Cardano.Api.LedgerState.Core
+import           Cardano.Api.LedgerState.Fold
