+++ cardano-api/internal/Cardano/Api/LedgerState/Core.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleContexts #-}
@@ -34,20 +33,10 @@ module Cardano.Api.LedgerState.Core
   , applyBlockWithEvents
   , AnyNewEpochState(..)
   , getAnyNewEpochState
-
-    -- * Traversing the block chain
-  , foldBlocks
-  , FoldStatus(..)
-  , chainSyncClientWithLedgerState
-  , chainSyncClientPipelinedWithLedgerState
-
-   -- * Ledger state conditions
-  , LedgerStateCondition(..)
-  , foldEpochState
+  , getNewEpochState
 
    -- * Errors
   , LedgerStateError(..)
-  , FoldBlocksError(..)
   , GenesisConfigError(..)
   , InitialLedgerStateError(..)
 
@@ -58,6 +47,7 @@ module Cardano.Api.LedgerState.Core
   , nextEpochEligibleLeadershipSlots
   -- * Node Config
   , NodeConfig(..)
+  , parseNodeConfig
   -- ** Network Config
   , NodeConfigFile
   , readNodeConfig
@@ -82,27 +72,38 @@ module Cardano.Api.LedgerState.Core
   , Env(..)
   , genesisConfigToEnv
 
+  -- ** Ledger State Events
+  , LedgerStateEvents
+  , toLedgerStateEvents
+
+  -- ** Ledger State History
+  , History
+  , LedgerStateHistory
+  , extractHistory
+
+  -- ** Ledger State utilities
+  , getBlockNo
+  , pushLedgerState
+  , rollBackLedgerStateHist
+  , atTerminationEpoch
+
+  , mkAdjustPath
+
   )
   where
 import           Cardano.Api.Block
 import           Cardano.Api.Certificate
 import           Cardano.Api.Eon.ShelleyBasedEra
 import           Cardano.Api.Eras.Case
-import           Cardano.Api.Eras.Core (forEraMaybeEon)
 import           Cardano.Api.Error as Api
 import           Cardano.Api.Genesis
 import           Cardano.Api.IO
-import           Cardano.Api.IPC (ConsensusModeParams (..),
-                   LocalChainSyncClient (LocalChainSyncClientPipelined),
-                   LocalNodeClientProtocols (..), LocalNodeClientProtocolsInMode,
-                   LocalNodeConnectInfo (..), connectToLocalNode)
+import           Cardano.Api.IPC ()
 import           Cardano.Api.Keys.Praos
 import           Cardano.Api.LedgerEvents.ConvertLedgerEvent
 import           Cardano.Api.LedgerEvents.LedgerEvent
-import           Cardano.Api.Modes (EpochSlots (..))
 import qualified Cardano.Api.Modes as Api
 import           Cardano.Api.Monad.Error
-import           Cardano.Api.NetworkId (NetworkId (..), NetworkMagic (NetworkMagic))
 import           Cardano.Api.Pretty
 import           Cardano.Api.Query (CurrentEpochState (..), PoolDistribution (unPoolDistr),
                    ProtocolState, SerialisedCurrentEpochState (..), SerialisedPoolDistribution,
@@ -114,7 +115,7 @@ import           Cardano.Api.Utils (textShow)
 import qualified Cardano.Binary as CBOR
 import qualified Cardano.Chain.Genesis
 import qualified Cardano.Chain.Update
-import           Cardano.Crypto (ProtocolMagicId (unProtocolMagicId), RequiresNetworkMagic (..))
+import           Cardano.Crypto (RequiresNetworkMagic (..))
 import qualified Cardano.Crypto.Hash.Blake2b
 import qualified Cardano.Crypto.Hash.Class
 import qualified Cardano.Crypto.Hashing
@@ -139,7 +140,7 @@ import           Cardano.Protocol.TPraos.BHeader (checkLeaderNatValue)
 import qualified Cardano.Protocol.TPraos.BHeader as TPraos
 import           Cardano.Slotting.EpochInfo (EpochInfo)
 import qualified Cardano.Slotting.EpochInfo.API as Slot
-import           Cardano.Slotting.Slot (WithOrigin (At, Origin))
+import           Cardano.Slotting.Slot (WithOrigin (..))
 import qualified Cardano.Slotting.Slot as Slot
 import qualified Ouroboros.Consensus.Block.Abstract as Consensus
 import           Ouroboros.Consensus.Block.Forging (BlockForging)
@@ -150,7 +151,6 @@ import qualified Ouroboros.Consensus.Cardano.Block as Consensus
 import qualified Ouroboros.Consensus.Cardano.CanHardFork as Consensus
 import qualified Ouroboros.Consensus.Cardano.Node as Consensus
 import qualified Ouroboros.Consensus.Config as Consensus
-import qualified Ouroboros.Consensus.HardFork.Combinator as Consensus
 import qualified Ouroboros.Consensus.HardFork.Combinator.AcrossEras as HFC
 import qualified Ouroboros.Consensus.HardFork.Combinator.Basics as HFC
 import qualified Ouroboros.Consensus.HardFork.Combinator.Serialisation.Common as HFC
@@ -171,16 +171,10 @@ import qualified Ouroboros.Consensus.Shelley.Ledger.Ledger as Shelley
 import           Ouroboros.Consensus.TypeFamilyWrappers (WrapLedgerEvent (WrapLedgerEvent))
 import           Ouroboros.Network.Block (blockNo)
 import qualified Ouroboros.Network.Block
-import qualified Ouroboros.Network.Protocol.ChainSync.Client as CS
-import qualified Ouroboros.Network.Protocol.ChainSync.ClientPipelined as CSP
-import           Ouroboros.Network.Protocol.ChainSync.PipelineDecision
 
-import           Control.Concurrent
-import           Control.DeepSeq
 import           Control.Error.Util (note)
 import           Control.Exception.Safe
 import           Control.Monad
-import           Control.Monad.State.Strict
 import           Data.Aeson as Aeson (FromJSON (parseJSON), Object, eitherDecodeStrict', withObject,
                    (.:), (.:?))
 import           Data.Aeson.Types (Parser)
@@ -193,7 +187,6 @@ import qualified Data.ByteString.Base16 as Base16
 import qualified Data.ByteString.Lazy as LB
 import           Data.ByteString.Short as BSS
 import           Data.Foldable
-import           Data.IORef
 import qualified Data.List as List
 import           Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
@@ -216,7 +209,6 @@ import           Data.Word
 import qualified Data.Yaml as Yaml
 import           Formatting.Buildable (build)
 import           Lens.Micro
-import           Network.TypedProtocol.Pipelined (Nat (..))
 import           System.FilePath
 
 data InitialLedgerStateError
@@ -367,451 +359,6 @@ pattern LedgerStateConway st <- LedgerState  (Consensus.LedgerStateConway st)
            , LedgerStateAlonzo
            , LedgerStateBabbage
            , LedgerStateConway #-}
-
-data FoldBlocksError
-  = FoldBlocksInitialLedgerStateError !InitialLedgerStateError
-  | FoldBlocksApplyBlockError !LedgerStateError
-  | FoldBlocksIOException !IOException
-  deriving Show
-
-instance Error FoldBlocksError where
-  prettyError = \case
-    FoldBlocksInitialLedgerStateError err -> prettyError err
-    FoldBlocksApplyBlockError err -> "Failed when applying a block:" <+> prettyError err
-    FoldBlocksIOException err -> "IOException:" <+> prettyException err
-
--- | Type that lets us decide whether to continue or stop
--- the fold from within our accumulation function.
-data FoldStatus
-  = ContinueFold
-  | StopFold
-  | DebugFold
-  deriving (Show, Eq)
-
--- | Monadic fold over all blocks and ledger states. Stopping @k@ blocks before
--- the node's tip where @k@ is the security parameter.
-foldBlocks
-  :: forall a t m. ()
-  => Show a
-  => MonadIOTransError FoldBlocksError t m
-  => NodeConfigFile 'In
-  -- ^ Path to the cardano-node config file (e.g. <path to cardano-node project>/configuration/cardano/mainnet-config.json)
-  -> SocketPath
-  -- ^ Path to local cardano-node socket. This is the path specified by the @--socket-path@ command line option when running the node.
-  -> ValidationMode
-  -> a
-  -- ^ The initial accumulator state.
-  -> (Env -> LedgerState -> [LedgerEvent] -> BlockInMode -> a -> IO (a, FoldStatus))
-  -- ^ Accumulator function Takes:
-  --
-  --  * Environment (this is a constant over the whole fold).
-  --  * The Ledger state (with block @i@ applied) at block @i@.
-  --  * The Ledger events resulting from applying block @i@.
-  --  * Block @i@.
-  --  * The accumulator state at block @i - 1@.
-  --
-  -- And returns:
-  --
-  --  * The accumulator state at block @i@
-  --  * A type indicating whether to stop or continue folding.
-  --
-  -- Note: This function can safely assume no rollback will occur even though
-  -- internally this is implemented with a client protocol that may require
-  -- rollback. This is achieved by only calling the accumulator on states/blocks
-  -- that are older than the security parameter, k. This has the side effect of
-  -- truncating the last k blocks before the node's tip.
-  -> t m a
-  -- ^ The final state
-foldBlocks nodeConfigFilePath socketPath validationMode state0 accumulate = handleIOExceptions $ do
-  -- NOTE this was originally implemented with a non-pipelined client then
-  -- changed to a pipelined client for a modest speedup:
-  --  * Non-pipelined: 1h  0m  19s
-  --  * Pipelined:        46m  23s
-
-  (env, ledgerState) <- modifyError FoldBlocksInitialLedgerStateError $ initialLedgerState nodeConfigFilePath
-
-  -- Place to store the accumulated state
-  -- This is a bit ugly, but easy.
-  errorIORef <- liftIO $ newIORef Nothing
-  stateIORef <- liftIO $ newIORef state0
-
-  -- Derive the NetworkId as described in network-magic.md from the
-  -- cardano-ledger-specs repo.
-  let byronConfig
-        = (\(Consensus.WrapPartialLedgerConfig (Consensus.ByronPartialLedgerConfig bc _) :* _) -> bc)
-        . HFC.getPerEraLedgerConfig
-        . HFC.hardForkLedgerConfigPerEra
-        $ envLedgerConfig env
-
-      networkMagic
-        = NetworkMagic
-        $ unProtocolMagicId
-        $ Cardano.Chain.Genesis.gdProtocolMagicId
-        $ Cardano.Chain.Genesis.configGenesisData byronConfig
-
-      networkId' = case Cardano.Chain.Genesis.configReqNetMagic byronConfig of
-        RequiresNoMagic -> Mainnet
-        RequiresMagic -> Testnet networkMagic
-
-      cardanoModeParams = CardanoModeParams . EpochSlots $ 10 * envSecurityParam env
-
-  -- Connect to the node.
-  let connectInfo = LocalNodeConnectInfo
-        { localConsensusModeParams  = cardanoModeParams
-        , localNodeNetworkId        = networkId'
-        , localNodeSocketPath       = socketPath
-        }
-
-  liftIO $ connectToLocalNode
-    connectInfo
-    (protocols stateIORef errorIORef env ledgerState)
-
-  liftIO (readIORef errorIORef) >>= \case
-    Just err -> throwError (FoldBlocksApplyBlockError err)
-    Nothing -> liftIO $ readIORef stateIORef
-  where
-    protocols :: ()
-      => IORef a
-      -> IORef (Maybe LedgerStateError)
-      -> Env
-      -> LedgerState
-      -> LocalNodeClientProtocolsInMode
-    protocols stateIORef errorIORef env ledgerState =
-        LocalNodeClientProtocols {
-          localChainSyncClient    = LocalChainSyncClientPipelined (chainSyncClient 50 stateIORef errorIORef env ledgerState),
-          localTxSubmissionClient = Nothing,
-          localStateQueryClient   = Nothing,
-          localTxMonitoringClient = Nothing
-        }
-
-    -- | Defines the client side of the chain sync protocol.
-    chainSyncClient :: Word16
-                    -- ^ The maximum number of concurrent requests.
-                    -> IORef a
-                    -- ^ State accumulator. Written to on every block.
-                    -> IORef (Maybe LedgerStateError)
-                    -- ^ Resulting error if any. Written to once on protocol
-                    -- completion.
-                    -> Env
-                    -> LedgerState
-                    -> CSP.ChainSyncClientPipelined
-                        BlockInMode
-                        ChainPoint
-                        ChainTip
-                        IO ()
-                    -- ^ Client returns maybe an error.
-    chainSyncClient pipelineSize stateIORef errorIORef env ledgerState0
-      = CSP.ChainSyncClientPipelined $ pure $ clientIdle_RequestMoreN Origin Origin Zero initialLedgerStateHistory
-      where
-          initialLedgerStateHistory = Seq.singleton (0, (ledgerState0, []), Origin)
-
-          clientIdle_RequestMoreN
-            :: WithOrigin BlockNo
-            -> WithOrigin BlockNo
-            -> Nat n -- Number of requests inflight.
-            -> LedgerStateHistory
-            -> CSP.ClientPipelinedStIdle n BlockInMode ChainPoint ChainTip IO ()
-          clientIdle_RequestMoreN clientTip serverTip n knownLedgerStates
-            = case pipelineDecisionMax pipelineSize n clientTip serverTip  of
-                Collect -> case n of
-                  Succ predN -> CSP.CollectResponse Nothing (clientNextN predN knownLedgerStates)
-                _ -> CSP.SendMsgRequestNextPipelined (pure ()) (clientIdle_RequestMoreN clientTip serverTip (Succ n) knownLedgerStates)
-
-          clientNextN
-            :: Nat n -- Number of requests inflight.
-            -> LedgerStateHistory
-            -> CSP.ClientStNext n BlockInMode ChainPoint ChainTip IO ()
-          clientNextN n knownLedgerStates =
-            CSP.ClientStNext {
-                CSP.recvMsgRollForward = \blockInMode@(BlockInMode _ block@(Block (BlockHeader slotNo _ currBlockNo) _)) serverChainTip -> do
-                  let newLedgerStateE = applyBlock
-                        env
-                        (maybe
-                          (error "Impossible! Missing Ledger state")
-                          (\(_,(ledgerState, _),_) -> ledgerState)
-                          (Seq.lookup 0 knownLedgerStates)
-                        )
-                        validationMode
-                        block
-                  case newLedgerStateE of
-                    Left err -> clientIdle_DoneNwithMaybeError n (Just err)
-                    Right newLedgerState -> do
-                      let (knownLedgerStates', committedStates) = pushLedgerState env knownLedgerStates slotNo newLedgerState blockInMode
-                          newClientTip = At currBlockNo
-                          newServerTip = fromChainTip serverChainTip
-
-                          ledgerStateSingleFold
-                            :: (SlotNo, (LedgerState, [LedgerEvent]), WithOrigin BlockInMode) -- Ledger events for a single block
-                            -> IO FoldStatus
-                          ledgerStateSingleFold (_, _, Origin) = return ContinueFold
-                          ledgerStateSingleFold (_, (ledgerState, ledgerEvents), At currBlock) = do
-                            accumulatorState <- readIORef stateIORef
-                            (newState, foldStatus) <- accumulate
-                                                      env
-                                                      ledgerState
-                                                      ledgerEvents
-                                                      currBlock
-                                                      accumulatorState
-                            atomicWriteIORef stateIORef newState
-                            return foldStatus
-
-                          ledgerStateRecurser
-                            :: Seq (SlotNo, LedgerStateEvents, WithOrigin BlockInMode) -- Ledger events for multiple blocks
-                            -> IO FoldStatus
-                          ledgerStateRecurser states = go (toList states) ContinueFold
-                            where
-                              go [] foldStatus = return foldStatus
-                              go (s : rest) ContinueFold = do
-                                newFoldStatus <- ledgerStateSingleFold s
-                                go rest newFoldStatus
-                              go _          StopFold    = go [] StopFold
-                              go _          DebugFold   = go [] DebugFold
-
-                      -- NB: knownLedgerStates' is the new ledger state history i.e k blocks from the tip
-                      -- or also known as the mutable blocks. We default to using the mutable blocks.
-                      finalFoldStatus <- ledgerStateRecurser knownLedgerStates'
-
-                      case finalFoldStatus of
-                        StopFold ->
-                          -- We return StopFold in our accumulate function if we want to terminate the fold.
-                          -- This allow us to check for a specific condition in our accumulate function
-                          -- and then terminate e.g a specific stake pool was registered
-                          let noError = Nothing
-                          in clientIdle_DoneNwithMaybeError n noError
-
-                        DebugFold -> do
-                          currentIORefState <- readIORef stateIORef
-
-                          -- Useful for debugging:
-                          let !ioRefErr = DebugError . force
-                                            $ unlines [ "newClientTip: " <> show newClientTip
-                                                      , "newServerTip: " <> show newServerTip
-                                                      , "newLedgerState: " <> show (snd newLedgerState)
-                                                      , "knownLedgerStates: " <> show (extractHistory knownLedgerStates)
-                                                      , "committedStates: " <> show (extractHistory committedStates)
-                                                      , "numberOfRequestsInFlight: " <> show n
-                                                      , "k: " <> show (envSecurityParam env)
-                                                      , "Current IORef State: " <> show currentIORefState
-                                                      ]
-                          clientIdle_DoneNwithMaybeError n $ Just ioRefErr
-
-                        ContinueFold -> return $ clientIdle_RequestMoreN newClientTip newServerTip n knownLedgerStates'
-
-              , CSP.recvMsgRollBackward = \chainPoint serverChainTip -> do
-                  let newClientTip = Origin -- We don't actually keep track of blocks so we temporarily "forget" the tip.
-                      newServerTip = fromChainTip serverChainTip
-                      truncatedKnownLedgerStates = case chainPoint of
-                          ChainPointAtGenesis -> initialLedgerStateHistory
-                          ChainPoint slotNo _ -> rollBackLedgerStateHist knownLedgerStates slotNo
-                  return (clientIdle_RequestMoreN newClientTip newServerTip n truncatedKnownLedgerStates)
-              }
-
-
-          clientIdle_DoneNwithMaybeError
-            :: Nat n -- Number of requests inflight.
-            -> Maybe LedgerStateError -- Return value (maybe an error)
-            -> IO (CSP.ClientPipelinedStIdle n BlockInMode ChainPoint ChainTip IO ())
-          clientIdle_DoneNwithMaybeError n errorMay = case n of
-            Succ predN -> return (CSP.CollectResponse Nothing (clientNext_DoneNwithMaybeError predN errorMay)) -- Ignore remaining message responses
-            Zero -> do
-              writeIORef errorIORef errorMay
-              return (CSP.SendMsgDone ())
-
-          clientNext_DoneNwithMaybeError
-            :: Nat n -- Number of requests inflight.
-            -> Maybe LedgerStateError -- Return value (maybe an error)
-            -> CSP.ClientStNext n BlockInMode ChainPoint ChainTip IO ()
-          clientNext_DoneNwithMaybeError n errorMay =
-            CSP.ClientStNext {
-                CSP.recvMsgRollForward = \_ _ -> clientIdle_DoneNwithMaybeError n errorMay
-              , CSP.recvMsgRollBackward = \_ _ -> clientIdle_DoneNwithMaybeError n errorMay
-              }
-
-          fromChainTip :: ChainTip -> WithOrigin BlockNo
-          fromChainTip ct = case ct of
-            ChainTipAtGenesis -> Origin
-            ChainTip _ _ bno -> At bno
-
-
--- | Wrap a 'ChainSyncClient' with logic that tracks the ledger state.
-chainSyncClientWithLedgerState
-  :: forall m a.
-     Monad m
-  => Env
-  -> LedgerState
-  -- ^ Initial ledger state
-  -> ValidationMode
-  -> CS.ChainSyncClient (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent]))
-                        ChainPoint
-                        ChainTip
-                        m
-                        a
-  -- ^ A client to wrap. The block is annotated with a 'Either LedgerStateError
-  -- LedgerState'. This is either an error from validating a block or
-  -- the current 'LedgerState' from applying the current block. If we
-  -- trust the node, then we generally expect blocks to validate. Also note that
-  -- after a block fails to validate we may still roll back to a validated
-  -- block, in which case the valid 'LedgerState' will be passed here again.
-  -> CS.ChainSyncClient BlockInMode
-                        ChainPoint
-                        ChainTip
-                        m
-                        a
-  -- ^ A client that acts just like the wrapped client but doesn't require the
-  -- 'LedgerState' annotation on the block type.
-chainSyncClientWithLedgerState env ledgerState0 validationMode (CS.ChainSyncClient clientTop)
-  = CS.ChainSyncClient (goClientStIdle (Right initialLedgerStateHistory) <$> clientTop)
-  where
-    goClientStIdle
-      :: Either LedgerStateError (History (Either LedgerStateError LedgerStateEvents))
-      -> CS.ClientStIdle (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent])) ChainPoint ChainTip m a
-      -> CS.ClientStIdle  BlockInMode                                                        ChainPoint ChainTip m a
-    goClientStIdle history client = case client of
-      CS.SendMsgRequestNext a b -> CS.SendMsgRequestNext a (goClientStNext history b)
-      CS.SendMsgFindIntersect ps a -> CS.SendMsgFindIntersect ps (goClientStIntersect history a)
-      CS.SendMsgDone a -> CS.SendMsgDone a
-
-    -- This is where the magic happens. We intercept the blocks and rollbacks
-    -- and use it to maintain the correct ledger state.
-    goClientStNext
-      :: Either LedgerStateError (History (Either LedgerStateError LedgerStateEvents))
-      -> CS.ClientStNext (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent])) ChainPoint ChainTip m a
-      -> CS.ClientStNext  BlockInMode                                                        ChainPoint ChainTip m a
-    goClientStNext (Left err) (CS.ClientStNext recvMsgRollForward recvMsgRollBackward) = CS.ClientStNext
-      (\blkInMode tip -> CS.ChainSyncClient $
-            goClientStIdle (Left err) <$> CS.runChainSyncClient
-                (recvMsgRollForward (blkInMode, Left err) tip)
-      )
-      (\point tip -> CS.ChainSyncClient $
-            goClientStIdle (Left err) <$> CS.runChainSyncClient (recvMsgRollBackward point tip)
-      )
-    goClientStNext (Right history) (CS.ClientStNext recvMsgRollForward recvMsgRollBackward) = CS.ClientStNext
-      (\blkInMode@(BlockInMode _ blk@(Block (BlockHeader slotNo _ _) _)) tip -> CS.ChainSyncClient $ let
-          newLedgerStateE = case Seq.lookup 0 history of
-            Nothing -> error "Impossible! History should always be non-empty"
-            Just (_, Left err, _) -> Left err
-            Just (_, Right (oldLedgerState, _), _) -> applyBlock
-                  env
-                  oldLedgerState
-                  validationMode
-                  blk
-          (history', _) = pushLedgerState env history slotNo newLedgerStateE blkInMode
-          in goClientStIdle (Right history') <$> CS.runChainSyncClient
-                (recvMsgRollForward (blkInMode, newLedgerStateE) tip)
-      )
-      (\point tip -> let
-          oldestSlot = case history of
-            _ Seq.:|> (s, _, _) -> s
-            Seq.Empty -> error "Impossible! History should always be non-empty"
-          history' = (\h -> if Seq.null h
-                              then Left (InvalidRollback oldestSlot point)
-                              else Right h)
-                  $ case point of
-                        ChainPointAtGenesis -> initialLedgerStateHistory
-                        ChainPoint slotNo _ -> rollBackLedgerStateHist history slotNo
-        in CS.ChainSyncClient $ goClientStIdle history' <$> CS.runChainSyncClient (recvMsgRollBackward point tip)
-      )
-
-    goClientStIntersect
-      :: Either LedgerStateError (History (Either LedgerStateError LedgerStateEvents))
-      -> CS.ClientStIntersect (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent])) ChainPoint ChainTip m a
-      -> CS.ClientStIntersect  BlockInMode                                                        ChainPoint ChainTip m a
-    goClientStIntersect history (CS.ClientStIntersect recvMsgIntersectFound recvMsgIntersectNotFound) = CS.ClientStIntersect
-      (\point tip -> CS.ChainSyncClient (goClientStIdle history <$> CS.runChainSyncClient (recvMsgIntersectFound point tip)))
-      (\tip -> CS.ChainSyncClient (goClientStIdle history <$> CS.runChainSyncClient (recvMsgIntersectNotFound tip)))
-
-    initialLedgerStateHistory :: History (Either LedgerStateError LedgerStateEvents)
-    initialLedgerStateHistory = Seq.singleton (0, Right (ledgerState0, []), Origin)
-
--- | See 'chainSyncClientWithLedgerState'.
-chainSyncClientPipelinedWithLedgerState
-  :: forall m a.
-     Monad m
-  => Env
-  -> LedgerState
-  -> ValidationMode
-  -> CSP.ChainSyncClientPipelined
-                        (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent]))
-                        ChainPoint
-                        ChainTip
-                        m
-                        a
-  -> CSP.ChainSyncClientPipelined
-                        BlockInMode
-                        ChainPoint
-                        ChainTip
-                        m
-                        a
-chainSyncClientPipelinedWithLedgerState env ledgerState0 validationMode (CSP.ChainSyncClientPipelined clientTop)
-  = CSP.ChainSyncClientPipelined (goClientPipelinedStIdle (Right initialLedgerStateHistory) Zero <$> clientTop)
-  where
-    goClientPipelinedStIdle
-      :: Either LedgerStateError (History (Either LedgerStateError LedgerStateEvents))
-      -> Nat n
-      -> CSP.ClientPipelinedStIdle n (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent])) ChainPoint ChainTip m a
-      -> CSP.ClientPipelinedStIdle n  BlockInMode                                                        ChainPoint ChainTip m a
-    goClientPipelinedStIdle history n client = case client of
-      CSP.SendMsgRequestNext a b -> CSP.SendMsgRequestNext a (goClientStNext history n b)
-      CSP.SendMsgRequestNextPipelined m a ->  CSP.SendMsgRequestNextPipelined m (goClientPipelinedStIdle history (Succ n) a)
-      CSP.SendMsgFindIntersect ps a -> CSP.SendMsgFindIntersect ps (goClientPipelinedStIntersect history n a)
-      CSP.CollectResponse a b -> case n of
-        Succ nPrev -> CSP.CollectResponse ((fmap . fmap) (goClientPipelinedStIdle history n) a) (goClientStNext history nPrev b)
-      CSP.SendMsgDone a -> CSP.SendMsgDone a
-
-    -- This is where the magic happens. We intercept the blocks and rollbacks
-    -- and use it to maintain the correct ledger state.
-    goClientStNext
-      :: Either LedgerStateError (History (Either LedgerStateError LedgerStateEvents))
-      -> Nat n
-      -> CSP.ClientStNext n (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent])) ChainPoint ChainTip m a
-      -> CSP.ClientStNext n  BlockInMode                                                        ChainPoint ChainTip m a
-    goClientStNext (Left err) n (CSP.ClientStNext recvMsgRollForward recvMsgRollBackward) = CSP.ClientStNext
-      (\blkInMode tip ->
-          goClientPipelinedStIdle (Left err) n <$> recvMsgRollForward
-            (blkInMode, Left err) tip
-      )
-      (\point tip ->
-          goClientPipelinedStIdle (Left err) n <$> recvMsgRollBackward point tip
-      )
-    goClientStNext (Right history) n (CSP.ClientStNext recvMsgRollForward recvMsgRollBackward) = CSP.ClientStNext
-      (\blkInMode@(BlockInMode _ blk@(Block (BlockHeader slotNo _ _) _)) tip -> let
-          newLedgerStateE = case Seq.lookup 0 history of
-            Nothing -> error "Impossible! History should always be non-empty"
-            Just (_, Left err, _) -> Left err
-            Just (_, Right (oldLedgerState, _), _) -> applyBlock
-                  env
-                  oldLedgerState
-                  validationMode
-                  blk
-          (history', _) = pushLedgerState env history slotNo newLedgerStateE blkInMode
-        in goClientPipelinedStIdle (Right history') n <$> recvMsgRollForward
-              (blkInMode, newLedgerStateE) tip
-      )
-      (\point tip -> let
-          oldestSlot = case history of
-            _ Seq.:|> (s, _, _) -> s
-            Seq.Empty -> error "Impossible! History should always be non-empty"
-          history' = (\h -> if Seq.null h
-                              then Left (InvalidRollback oldestSlot point)
-                              else Right h)
-                  $ case point of
-                        ChainPointAtGenesis -> initialLedgerStateHistory
-                        ChainPoint slotNo _ -> rollBackLedgerStateHist history slotNo
-        in goClientPipelinedStIdle history' n <$> recvMsgRollBackward point tip
-      )
-
-    goClientPipelinedStIntersect
-      :: Either LedgerStateError (History (Either LedgerStateError LedgerStateEvents))
-      -> Nat n
-      -> CSP.ClientPipelinedStIntersect (BlockInMode, Either LedgerStateError (LedgerState, [LedgerEvent])) ChainPoint ChainTip m a
-      -> CSP.ClientPipelinedStIntersect  BlockInMode                                                        ChainPoint ChainTip m a
-    goClientPipelinedStIntersect history _ (CSP.ClientPipelinedStIntersect recvMsgIntersectFound recvMsgIntersectNotFound) = CSP.ClientPipelinedStIntersect
-      (\point tip -> goClientPipelinedStIdle history Zero <$> recvMsgIntersectFound point tip)
-      (\tip -> goClientPipelinedStIdle history Zero <$> recvMsgIntersectNotFound tip)
-
-    initialLedgerStateHistory :: History (Either LedgerStateError LedgerStateEvents)
-    initialLedgerStateHistory = Seq.singleton (0, Right (ledgerState0, []), Origin)
-
 
 extractHistory
   :: History LedgerStateEvents
@@ -1843,11 +1390,6 @@ constructGlobals sGen eInfo (Ledger.ProtVer majorPParamsVer _) =
 
 --------------------------------------------------------------------------
 
-data LedgerStateCondition
-  = ConditionMet
-  | ConditionNotMet
-  deriving (Show, Eq)
-
 data AnyNewEpochState where
   AnyNewEpochState
     :: ShelleyBasedEra era
@@ -1858,227 +1400,6 @@ instance Show AnyNewEpochState where
   showsPrec p (AnyNewEpochState sbe ledgerNewEpochState) =
     shelleyBasedEraConstraints sbe $ showsPrec p ledgerNewEpochState
 
-
--- | Reconstructs the ledger's new epoch state and applies a supplied condition to it for every block. This
--- function only terminates if the condition is met or we have reached the termination epoch. We need to
--- provide a termination epoch otherwise blocks would be applied indefinitely.
-foldEpochState
-  :: forall t m s. MonadIOTransError FoldBlocksError t m
-  => NodeConfigFile 'In
-  -- ^ Path to the cardano-node config file (e.g. <path to cardano-node project>/configuration/cardano/mainnet-config.json)
-  -> SocketPath
-  -- ^ Path to local cardano-node socket. This is the path specified by the @--socket-path@ command line option when running the node.
-  -> ValidationMode
-  -> EpochNo
-  -- ^ Termination epoch
-  -> s
-  -- ^ an initial value for the condition state
-  -> ( AnyNewEpochState
-         -> SlotNo
-         -> BlockNo
-         -> StateT s IO LedgerStateCondition
-     )
-  -- ^ Condition you want to check against the new epoch state.
-  --
-  --   'SlotNo' - Current (not necessarily the tip) slot number
-  --
-  --   'BlockNo' - Current (not necessarily the tip) block number
-  --
-  -- Note: This function can safely assume no rollback will occur even though
-  -- internally this is implemented with a client protocol that may require
-  -- rollback. This is achieved by only calling the accumulator on states/blocks
-  -- that are older than the security parameter, k. This has the side effect of
-  -- truncating the last k blocks before the node's tip.
-  -> t m (LedgerStateCondition, s)
-  -- ^ The final state
-foldEpochState nodeConfigFilePath socketPath validationMode terminationEpoch initialResult checkCondition = handleIOExceptions $ do
-  -- NOTE this was originally implemented with a non-pipelined client then
-  -- changed to a pipelined client for a modest speedup:
-  --  * Non-pipelined: 1h  0m  19s
-  --  * Pipelined:        46m  23s
-
-  (env, ledgerState) <- modifyError FoldBlocksInitialLedgerStateError
-                          $ initialLedgerState nodeConfigFilePath
-
-  -- Place to store the accumulated state
-  -- This is a bit ugly, but easy.
-  errorIORef <- modifyError FoldBlocksIOException . liftIO $ newIORef Nothing
-  -- This needs to be a full MVar by default. It serves as a mutual exclusion lock when executing
-  -- 'checkCondition' to ensure that states 's' are processed in order. This is assured by MVar fairness.
-  stateMv <- modifyError FoldBlocksIOException . liftIO $ newMVar (ConditionNotMet, initialResult)
-
-  -- Derive the NetworkId as described in network-magic.md from the
-  -- cardano-ledger-specs repo.
-  let byronConfig
-        = (\(Consensus.WrapPartialLedgerConfig (Consensus.ByronPartialLedgerConfig bc _) :* _) -> bc)
-        . HFC.getPerEraLedgerConfig
-        . HFC.hardForkLedgerConfigPerEra
-        $ envLedgerConfig env
-
-      networkMagic
-        = NetworkMagic
-        $ unProtocolMagicId
-        $ Cardano.Chain.Genesis.gdProtocolMagicId
-        $ Cardano.Chain.Genesis.configGenesisData byronConfig
-
-      networkId' = case Cardano.Chain.Genesis.configReqNetMagic byronConfig of
-        RequiresNoMagic -> Mainnet
-        RequiresMagic -> Testnet networkMagic
-
-      cardanoModeParams = CardanoModeParams . EpochSlots $ 10 * envSecurityParam env
-
-  -- Connect to the node.
-  let connectInfo = LocalNodeConnectInfo
-        { localConsensusModeParams  = cardanoModeParams
-        , localNodeNetworkId        = networkId'
-        , localNodeSocketPath       = socketPath
-        }
-
-  modifyError FoldBlocksIOException $ liftIO $ connectToLocalNode
-    connectInfo
-    (protocols stateMv errorIORef env ledgerState)
-
-  liftIO (readIORef errorIORef) >>= \case
-    Just err -> throwError $ FoldBlocksApplyBlockError err
-    Nothing -> modifyError FoldBlocksIOException . liftIO $ readMVar stateMv
-  where
-    protocols :: ()
-      => MVar (LedgerStateCondition, s)
-      -> IORef (Maybe LedgerStateError)
-      -> Env
-      -> LedgerState
-      -> LocalNodeClientProtocolsInMode
-    protocols stateMv errorIORef env ledgerState =
-        LocalNodeClientProtocols {
-          localChainSyncClient    = LocalChainSyncClientPipelined (chainSyncClient 50 stateMv errorIORef env ledgerState),
-          localTxSubmissionClient = Nothing,
-          localStateQueryClient   = Nothing,
-          localTxMonitoringClient = Nothing
-        }
-
-    -- | Defines the client side of the chain sync protocol.
-    chainSyncClient :: Word16
-                    -- ^ The maximum number of concurrent requests.
-                    -> MVar (LedgerStateCondition, s)
-                    -- ^ State accumulator. Written to on every block.
-                    -> IORef (Maybe LedgerStateError)
-                    -- ^ Resulting error if any. Written to once on protocol
-                    -- completion.
-                    -> Env
-                    -> LedgerState
-                    -> CSP.ChainSyncClientPipelined
-                        BlockInMode
-                        ChainPoint
-                        ChainTip
-                        IO ()
-                    -- ^ Client returns maybe an error.
-    chainSyncClient pipelineSize stateMv errorIORef' env ledgerState0
-      = CSP.ChainSyncClientPipelined $ pure $ clientIdle_RequestMoreN Origin Origin Zero initialLedgerStateHistory
-      where
-          initialLedgerStateHistory = Seq.singleton (0, (ledgerState0, []), Origin)
-
-          clientIdle_RequestMoreN
-            :: WithOrigin BlockNo
-            -> WithOrigin BlockNo
-            -> Nat n -- Number of requests inflight.
-            -> LedgerStateHistory
-            -> CSP.ClientPipelinedStIdle n BlockInMode ChainPoint ChainTip IO ()
-          clientIdle_RequestMoreN clientTip serverTip n knownLedgerStates
-            = case pipelineDecisionMax pipelineSize n clientTip serverTip  of
-                Collect -> case n of
-                  Succ predN -> CSP.CollectResponse Nothing (clientNextN predN knownLedgerStates)
-                _ -> CSP.SendMsgRequestNextPipelined (pure ()) (clientIdle_RequestMoreN clientTip serverTip (Succ n) knownLedgerStates)
-
-          clientNextN
-            :: Nat n -- Number of requests inflight.
-            -> LedgerStateHistory
-            -> CSP.ClientStNext n BlockInMode ChainPoint ChainTip IO ()
-          clientNextN n knownLedgerStates =
-            CSP.ClientStNext {
-                CSP.recvMsgRollForward = \blockInMode@(BlockInMode era block@(Block (BlockHeader slotNo _ currBlockNo) _)) serverChainTip -> do
-                  let newLedgerStateE = applyBlock
-                        env
-                        (maybe
-                          (error "Impossible! Missing Ledger state")
-                          (\(_,(ledgerState, _),_) -> ledgerState)
-                          (Seq.lookup 0 knownLedgerStates)
-                        )
-                        validationMode
-                        block
-                  case forEraMaybeEon era of
-                    Nothing -> let !err = Just ByronEraUnsupported
-                               in clientIdle_DoneNwithMaybeError n err
-                    Just sbe ->
-                      case newLedgerStateE of
-                        Left err -> clientIdle_DoneNwithMaybeError n (Just err)
-                        Right new@(newLedgerState, ledgerEvents) -> do
-                          let (knownLedgerStates', _) = pushLedgerState env knownLedgerStates slotNo new blockInMode
-                              newClientTip = At currBlockNo
-                              newServerTip = fromChainTip serverChainTip
-                          case getNewEpochState sbe $ clsState newLedgerState of
-                            Left e ->
-                              let !err = Just e
-                              in clientIdle_DoneNwithMaybeError n err
-                            Right lState -> do
-                              let newEpochState = AnyNewEpochState sbe lState
-                              -- Run the condition function in an exclusive lock.
-                              -- There can be only one place where `takeMVar stateMv` exists otherwise this
-                              -- code will deadlock!
-                              condition <- bracket (takeMVar stateMv) (tryPutMVar stateMv) $ \(_prevCondition, previousState) -> do
-                                updatedState@(!newCondition, !_) <- runStateT (checkCondition newEpochState slotNo currBlockNo) previousState
-                                putMVar stateMv updatedState
-                                pure newCondition
-                              -- Have we reached the termination epoch?
-                              case atTerminationEpoch terminationEpoch ledgerEvents of
-                                Just !currentEpoch -> do
-                                   -- confirmed this works: error $ "atTerminationEpoch: Terminated at: " <> show currentEpoch
-                                   let !err = Just $ TerminationEpochReached currentEpoch
-                                   clientIdle_DoneNwithMaybeError n err
-                                Nothing -> do
-                                  case condition of
-                                    ConditionMet ->
-                                       let !noError = Nothing
-                                       in clientIdle_DoneNwithMaybeError n noError
-                                    ConditionNotMet -> return $ clientIdle_RequestMoreN newClientTip newServerTip n knownLedgerStates'
-
-              , CSP.recvMsgRollBackward = \chainPoint serverChainTip -> do
-                  let newClientTip = Origin -- We don't actually keep track of blocks so we temporarily "forget" the tip.
-                      newServerTip = fromChainTip serverChainTip
-                      truncatedKnownLedgerStates = case chainPoint of
-                          ChainPointAtGenesis -> initialLedgerStateHistory
-                          ChainPoint slotNo _ -> rollBackLedgerStateHist knownLedgerStates slotNo
-                  return (clientIdle_RequestMoreN newClientTip newServerTip n truncatedKnownLedgerStates)
-              }
-
-
-
-          clientIdle_DoneNwithMaybeError
-            :: Nat n -- Number of requests inflight.
-            -> Maybe LedgerStateError -- Return value (maybe an error)
-            -> IO (CSP.ClientPipelinedStIdle n BlockInMode ChainPoint ChainTip IO ())
-          clientIdle_DoneNwithMaybeError n errorMay = case n of
-            Succ predN -> do
-              return (CSP.CollectResponse Nothing (clientNext_DoneNwithMaybeError predN errorMay)) -- Ignore remaining message responses
-            Zero -> do
-              atomicModifyIORef' errorIORef' . const $ (errorMay, ())
-              return (CSP.SendMsgDone ())
-
-          clientNext_DoneNwithMaybeError
-            :: Nat n -- Number of requests inflight.
-            -> Maybe LedgerStateError -- Return value (maybe an error)
-            -> CSP.ClientStNext n BlockInMode ChainPoint ChainTip IO ()
-          clientNext_DoneNwithMaybeError n errorMay =
-            CSP.ClientStNext {
-                CSP.recvMsgRollForward = \_ _ -> clientIdle_DoneNwithMaybeError n errorMay
-              , CSP.recvMsgRollBackward = \_ _ -> clientIdle_DoneNwithMaybeError n errorMay
-              }
-
-          fromChainTip :: ChainTip -> WithOrigin BlockNo
-          fromChainTip ct = case ct of
-            ChainTipAtGenesis -> Origin
-            ChainTip _ _ bno -> At bno
-
-
 atTerminationEpoch :: EpochNo -> [LedgerEvent] -> Maybe EpochNo
 atTerminationEpoch terminationEpoch events =
   listToMaybe
@@ -2086,6 +1407,3 @@ atTerminationEpoch terminationEpoch events =
     , let currentEpoch' = prdEpochNo poolReapDets
     , currentEpoch' >= terminationEpoch
     ]
-
-handleIOExceptions :: MonadIOTransError FoldBlocksError t m => ExceptT FoldBlocksError IO a -> t m a
-handleIOExceptions = liftEither <=< liftIO . fmap (join . first FoldBlocksIOException) . try . runExceptT
